requirements.txt
flask==2.3.2
pandas==2.1.0
scikit-learn==1.3.0
numpy==1.25.0
python-docx==0.8.11
pdfminer.six==20221105
pytest==7.4.0